channel-discovery/requirements.txt
# Channel Discovery Service Dependencies

# Web framework
fastapi>=0.104.0
uvicorn[standard]>=0.24.0

# Telegram
telethon>=1.34.0

# Database
sqlalchemy>=2.0.0

# Validation
pydantic>=2.0.0

# Templates
jinja2>=3.1.0

# Utilities
python-dotenv>=1.0.0
